fix(langy): live run card crashed on first render; move the SPA hook to hooks/
The card's poll policy closed over `sseConnected`, which was declared BELOW
the query it was passed to. TanStack evaluates `refetchInterval` synchronously
while `useQuery` runs, so every live simulation card threw
`ReferenceError: Cannot access 'sseConnected' before initialization` on its
first render. The drawer this card mirrors declares the SSE listener above the
query; the card had to invert that order to let the listener read the query's
`error` (the round-2 fix for holding a subscription open on a dead run).

Restores the drawer's order and carries the dead-run signal in state instead,
so the listener no longer needs a value the query has not produced yet.

The tests never saw it because the `~/utils/api` stub only returned `{ data }`
and never called `refetchInterval` — the one behaviour of the real client the
card depends on. The stub now calls it, and 6 of the file's 7 tests fail
against the old component with the exact ReferenceError.

Alongside, three placement fixes:

- `useSpaLinkClick` was the only hook under `logic/` while the feature keeps
  26 in `hooks/` (dev/docs/best_practices/react.md). Moved, imports rewritten.
- its unit test still named `spaLink`, a module that no longer exists; renamed
  to `isInternalHref.unit.test.ts` after the function it actually covers. It
  stays in the langy pool rather than moving next to `Markdown.tsx`: importing
  that component graph into the `src/components` pool is enough extra load to
  flake `OutputsSection`'s typing assertions (reproduced 4/4 runs, green 3/3
  without it).
- dropped an `eslint-disable` comment from `Markdown.tsx`; this repo lints
  with biome and has no eslint, so it read as a suppression that governs
  nothing.
